Many: more places where the job ID was an int/Integer

+++ src/alien/api/taskQueue/TaskQueueApiUtils.java
@@ -86,7 +86,7 @@ public class TaskQueueApiUtils {
 	 * @param bExpunge
 	 * @return a PS listing
 	 */
-	public List<Job> getMasterJobStatus(final int jobId, final Set<JobStatus> status, final List<Integer> id, final List<String> site, final boolean bPrintId, final boolean bPrintSite,
+	public List<Job> getMasterJobStatus(final long jobId, final Set<JobStatus> status, final List<Integer> id, final List<String> site, final boolean bPrintId, final boolean bPrintSite,
 			final boolean bMerge, final boolean bKill, final boolean bResubmit, final boolean bExpunge) {
 
 		try {
@@ -235,7 +235,7 @@ public class TaskQueueApiUtils {
 	 *
 	 * @return status of the kill
 	 */
-	public boolean killJob(final int queueId) {
+	public boolean killJob(final long queueId) {
 
 		try {
 			final KillJob j = new KillJob(commander.getUser(), commander.getRole(), queueId);
